+++ yazlab_2_frontend/Forms/Pages/GraphViewPage.cs
@@ -188,7 +188,7 @@ namespace yazlab_2_frontend.Forms.Pages
 
                 using var brush = new SolidBrush(selected ? Color.LightSkyBlue : n.NodeRengi);
                 using var borderPen = new Pen(selected ? Color.DodgerBlue : Color.DimGray, selected ? 3 : 2);
-                int NodeRadius = n.radius;
+                float NodeRadius = n.radius;
                 var rect = new RectangleF(p.X - NodeRadius, p.Y - NodeRadius, NodeRadius * 2, NodeRadius * 2);
                 e.Graphics.FillEllipse(brush, rect);
                 e.Graphics.DrawEllipse(borderPen, rect);
